refactor(useauth): removed useAuth, added SWR
I've decided to make a major refactor of my app using SWR instead of React Context. This is because
it provides everything I got from Context, but with more features. This will be a better structure
for my app going forward.

BREAKING CHANGE: Functionality may be broken and requires testing. Also getting a strange 404 error
that I can't figure out.