fix(sidebar context): fixed hydration mismatch
I got an error regarding hydration because of server rendering, isOpen would default to false
because of isBrowser() and then become true. I refactored it so that isBrowser is state and the
component only runs client side. Also, I refactored the isOpen  and collapse code on the sidebar
component. it was backwards before, and now it makes sense.