fix(navbar): sidebar state localstorage
Moved setting sidebar state to localstorage to the nav component so that it is only set when toggled
by user instead of every time isOpen changes. This makes localstorage and isOpen context not
synchronized. The way it works is that it defaults to not open, and checks localstorage. This is how
it was set up before but localstorage would change on re-render because of isOpen defaulting to
false and it would set it as false every time. Now, it keeps track of the user's preference, while
maintaining the default logic for the sidebar. It has the best of both worlds, a default state and a
user preference.