feat(files-accordion, project-helpers, form-wrapper): react-hook-form fields array for media
media is now passed with defaultValues and handled with react-hook-form fieldsarray. media can be
deleted which is passed with the submit form. Then, edit-project function runs after combined media
with any new uploaded media. upload-media-and-get-submit-data has been refactored to handle this.
before, media was received from a get request, now it is part of the form data.